apps/api-python/requirements.txt
fastapi
uvicorn
pdfplumber
requests
python-multipart 
supabase
python-dotenv
google-genai